requirements.txt
diff-match-patch==20121119
Django==2.0.2
django-import-export==0.7.0
et-xmlfile==1.0.1
jdcal==1.3
odfpy==1.3.6
openpyxl==2.5.0
pytz==2017.3
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
